refactor(spindle-mcp-server): update tools
