+++ packages/spindle-mcp-server/src/icon.ts
@@ -1,0 +1,43 @@
+import fs from 'fs';
+import path from 'path';
+
+interface IconInfo {
+  name: string;
+  path: string;
+  svgPath: string;
+  viewBox: string;
+}
+
+export async function getAllIcons(): Promise<string[]> {
+  const iconDir = path.join(__dirname, '../../spindle-ui/src/Icon');
+  const files = await fs.promises.readdir(iconDir);
+  return files
+    .filter((file) => file.endsWith('.tsx'))
+    .map((file) => file.replace('.tsx', ''));
+}
+
+function getIconPath(iconName: string): string {
+  return path.join(__dirname, '../../spindle-ui/src/Icon', `${iconName}.tsx`);
+}
+
+export async function getIconInfo(iconName: string): Promise<IconInfo | null> {
+  const iconPath = getIconPath(iconName);
+  if (!fs.existsSync(iconPath)) {
+    return null;
+  }
+
+  const content = await fs.promises.readFile(iconPath, 'utf-8');
+  const svgMatch = content.match(/viewBox="([^"]+)"/);
+  const pathMatch = content.match(/<path d="([^"]+)"/);
+
+  if (!svgMatch || !pathMatch) {
+    return null;
+  }
+
+  return {
+    name: iconName,
+    path: iconPath,
+    svgPath: pathMatch[1],
+    viewBox: svgMatch[1],
+  };
+}
